requirements.txt
psycopg2-binary==2.9.7
boto3==1.29.7
pandas==2.1.1
pyarrow==13.0.0
